Split the variant test suite into two translation units to manage memory usage by the compiler
